Split metrics array to its own file.
